# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Physics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes/{quizId}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizPublic"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No access to course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quizzes/{quizId}/attempts:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "201":
          description: Attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptResult"

  # ─── AI ──────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    QuizAnswerPublic:
      type: object
      required: [id, questionId, answerLabel, answerText]
      properties:
        id:
          type: integer
        questionId:
          type: integer
        answerLabel:
          type: string
        answerText:
          type: string
    QuizQuestionPublic:
      type: object
      required: [id, quizId, questionText, sortOrder, answers]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        sortOrder:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswerPublic"
    QuizPublic:
      type: object
      required: [id, topicId, title, questions]
      properties:
        id:
          type: integer
        topicId:
          type: integer
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestionPublic"
    QuizAttemptAnswerInput:
      type: object
      required: [questionId, selectedAnswerId]
      properties:
        questionId:
          type: integer
        selectedAnswerId:
          type: integer
    QuizAttemptInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAttemptAnswerInput"
    QuizAttemptAnswerResult:
      type: object
      required: [questionId, selectedAnswerId, isCorrect, correctAnswerId]
      properties:
        questionId:
          type: integer
        selectedAnswerId:
          type: integer
        isCorrect:
          type: boolean
        correctAnswerId:
          type: integer
    QuizAttemptResult:
      type: object
      required: [score, totalQuestions, percentage, passed, passThreshold, answers]
      properties:
        score:
          type: integer
        totalQuestions:
          type: integer
        percentage:
          type: integer
        passed:
          type: boolean
        passThreshold:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAttemptAnswerResult"
...